refactor: add useDisplayedCard
