+++ src/hooks/index.ts
@@ -1,2 +1,3 @@
 export { default as useDisplayedInput } from './useDisplayedInput';
 export { default as useInput } from './useInput';
+export { default as useDisplayedCard } from './useDisplayedCard';
