+++ src/pages/addCard/hooks/useCardNumberInput.ts
@@ -1,25 +1,12 @@
-import {
-	CARD_NUMBER_HYPHEN_INTERVAL,
-	CARD_NUMBER_NOT_MASKED_LENGTH,
-	CARN_NUMBER_LENGTH,
-} from '@constants';
-import { useDisplayedInput } from '@hooks';
-import { insertAtInterval, isNumber, replaceTo } from '@utils';
-import { useCallback } from 'react';
+import { CARN_NUMBER_LENGTH } from '@constants';
+import { useDisplayedCard, useDisplayedInput } from '@hooks';
+import { isNumber } from '@utils';
 
 export default function useCardNumberInput() {
-	const toMaskedCardNumber = useCallback((value: string) => {
-		const masked = replaceTo(
-			value,
-			'*',
-			CARD_NUMBER_NOT_MASKED_LENGTH,
-			value.length,
-		);
-		return insertAtInterval(masked, '-', CARD_NUMBER_HYPHEN_INTERVAL);
-	}, []);
+	const { toDisplayedCardNumber } = useDisplayedCard();
 
 	const { value, displayedValue, handleChange } = useDisplayedInput({
-		toDisplayed: toMaskedCardNumber,
+		toDisplayed: toDisplayedCardNumber,
 		maxLength: CARN_NUMBER_LENGTH,
 		validate: isNumber,
 	});
